requirements.txt
streamlit
opencv-python
numpy
tensorflow
pillow
streamlit-option-menu
